test_project/pyproject.toml
[project]
name = "Test Project"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "Tobi-De", email = "tobidegnon@proton.me" }
]
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.2",
    "fastapi-cli>=0.0.2",
    "uvicorn>=0.21.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "fastapi-users>=10.4.1",
    "fastapi-users-tortoise>=0.2.0",
    "tortoise-orm[asyncpg]>=0.19.3",
    "aerich>=0.7.1",
    "saq[hiredis,web]>=0.9.3",
    "hupper>=1.11",
    "python-multipart>=0.0.9",
    "sentry-sdk>=1.17.0",
    "rich>=13.3.2",
    "typer>=0.7.0",
    "watchgod>=0.8.2",
    "Jinja2>=3.1.2",
    "httpx>=0.23.3",
    "aioaws>=0.14",
    ]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.2",
    "factory-boy>=3.2.1",
    "pre-commit>=3.1.1",
    "mypy>=1.1.1",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "ipython>=8.11.0",
    "aiosmtpd>=1.4.4.post2",
    "honcho>=1.1.0",
]

[tool.aerich]
tortoise_orm = "test_project.db.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[tool.isort]
profile = "black"

[tool.mypy]
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
ignore_missing_imports = true

[tool.ruff]
line-length = 120
select = [
    "E", # https://github.com/charliermarsh/ruff#pycodestylet
    "F", # https://github.com/charliermarsh/ruff#pyflakes
    "UP", # https://github.com/charliermarsh/ruff#pyupgrade
    "N", # https://github.com/charliermarsh/ruff#pep8-naming
    "S", # https://github.com/charliermarsh/ruff#flake8-bandit
    "C", # https://github.com/charliermarsh/ruff#flake8-comprehensions
    "B", # https://github.com/charliermarsh/ruff#flake8-bugbear
    "T" # https://github.com/charliermarsh/ruff#flake8-print
]

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "test_project"
module-root = ""
